feat: add basic utils
